файлы сгруппированы по папкам, каждая структура вынесена в отдельный файл
