+++ FakeNFT/Scenes /Catalog/CatalogCollection/CatalogCollectionsProvider.swift
