+++ FakeNFT/Scenes /Catalog/Network/NftCollectionCover.swift
