+++ FakeNFT/Scenes /Catalog/NftCollection/NftCellModel.swift
