+++ FakeNFT/Scenes /Catalog/NftCollection/NftCollectionViewController.swift
